Added brk() and sbrk() systemcalls

+++ r3_kernel/src/system/posix/gettime.rs
@@ -7,7 +7,7 @@ const MONOTONIC_CLOCK: i32 = 1;
 pub fn sys_clock_gettime(
     clock_sel: i32,
     timeval_buffer: abi::UserAddress,
-) -> Result<i32, abi::Errno> {
+) -> Result<isize, abi::Errno> {
     let timeval = match clock_sel {
         // TODO: Handle time with unix epoch
         REALTIME_CLOCK => timer::PosixTimeval::from_ticks(),
